Moved fixtures/updated references
