+++ packages/sermon/src/featured-series.js
